ADAPT Timer: band-aid CurrentTicks to return UINT64. partial fixup d53ac3a97fb3f1a22d055c52e86b5eef0933eb50

+++ MicroFrameworkPK_v4_3/DeviceCode/Targets/Native/Krait/DeviceCode/Krait_TIMER/Krait__TIMER_Functions.cpp
@@ -100,8 +100,9 @@ UINT16 CPU_Timer_GetCounter(UINT16 Timer)
 	return (UINT16)g_Krait_Timer.GetCounter(Timer);
 }
 
-UINT32 CPU_Timer_CurrentTicks(UINT16 Timer)
+UINT64 CPU_Timer_CurrentTicks(UINT16 Timer)
 {
+#warning "######## FIXME CPU_Timer_CurrentTicks does not implement latest 64-bit API! ##############"
 	return g_Krait_Timer.GetCounter(0);
 }
